new Makefile-like execution model, reusing old code
