--- astropop/pipelines/specific/onam_polima2.py
@@ -1,1 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
